gioithieuvieclam/job_introduction/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.3
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-filter==24.1
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
django-perms==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==24.11.0
idna==3.7
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.5.2
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
